+++ src/extension.ts
@@ -4,18 +4,13 @@
 
 import * as vscode from 'vscode';
 import { workspace, ExtensionContext, commands, window, Selection, Range, Position } from 'vscode';
-import { FileJumper } from './fileJumper';
-import { ALCodeActionsProvider } from './alCodeActionsProvider';
-import { ALNavigatorExtensionContext } from './alNavigatorExtensionContext';
+import { FileJumper } from './fileJumper/fileJumper';
+import { ALCodeActionsProvider } from './CommandsActions/alCodeActionsProvider';
+//import { ALNavigatorExtensionContext } from './alNavigatorExtensionContext';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-	const alNavigatorExtensionContext = new ALNavigatorExtensionContext(context);
-	context.subscriptions.push(alNavigatorExtensionContext);
-	let alCodeActionsProvider : ALCodeActionsProvider = new ALCodeActionsProvider(alNavigatorExtensionContext);
-	context.subscriptions.push(vscode.languages.registerCodeActionsProvider('al', alCodeActionsProvider));
-	
 	let jumpToNextdataItemCmd = commands.registerCommand("extension.DataItem", () => {
 		FileJumper.jumpToNextDataItem();
 	});
@@ -35,14 +30,14 @@ export function activate(context: vscode.ExtensionContext) {
 	let jumpToActionsCommand = commands.registerCommand("extension.Actions", () => {
 		FileJumper.jumpToNextActions();
 	});
+	let alCodeActionsProvider : ALCodeActionsProvider = new ALCodeActionsProvider(context);
+	context.subscriptions.push(vscode.languages.registerCodeActionsProvider('al', alCodeActionsProvider));
 	
 	context.subscriptions.push(jumpToNextdataItemCmd);
 	context.subscriptions.push(jumpToOnAfterGetRecCmd);
 	context.subscriptions.push(jumpToNextTriggerCommand);
 	context.subscriptions.push(jumpToKeysCommand);
 	context.subscriptions.push(jumpToActionsCommand);
-
-    return alNavigatorExtensionContext;
 }
 
 // this method is called when your extension is deactivated